refactor(subscription): move "Subscription" decorator to a dedicated directory
