+++ tartiflette/__init__.py
@@ -1,3 +1,4 @@
 from tartiflette.resolver import Resolver, ResolverExecutorFactory
+from tartiflette.subscription import Subscription
 from tartiflette.sdl import build_graphql_schema_from_sdl
 from tartiflette.engine import Engine
