+++ tartiflette/subscription/__init__.py
@@ -1,0 +1,3 @@
+from .subscription import Subscription
+
+__all__ = ["Subscription"]
